core: Use transfer id for nbd server id and socket
When engine start the NBD server in vdsm during upload (e.g. restore) or
download (e.g. offline backup) it uses the image transfer ticket id as the
server id. This id is used to create a socket at /run/vdsm/nbd/{id} and for
stopping the server.

We have 2 UUIDs for an image transfer:
- The transfer id - the engine command id.
- The ticket id - used for authenticating the client with imageio server.

The ticket id is sensitive information that should not be logged while the
transfer is active. The transfer id is public information that may be logged
everywhere, and we want all events related to image transfer to have this id.

This change remove the ticket id from many logs in vdsm and imageio, and
replace the nbd socket /run/vdsm/nbd/{nbd-server-id} to use the transfer
id.

Change-Id: Ic6cfa674290b30e52f22549b06570906f06c1983
Bug-Url: https://bugzilla.redhat.com/2043283
Signed-off-by: Nir Soffer <nsoffer@redhat.com>